pyproject.toml
[project]
name = "fastapi-mlops"
version = "0.1.0"
description = "Basic MLOPS with fullstack"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.70",
    "catboost>=1.2.10",
    "dvc[s3]>=3.67.1",
    "fastapi>=0.135.3",
    "lightgbm>=4.6.0",
    "loguru>=0.7.3",
    "minio>=7.2.20",
    "mlflow>=3.10.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-decouple>=3.8",
    "xgboost>=3.2.0",
]
